Add RESTful api to view a questionnaire schema
For a given questionnaire we have added a view function
that generates a json representation of a questionnaire.

This will eventually be consumed by eq-survey-runner,
enabling the rendering of a questionnaire to a user.

This commit also adds support for django_debug_toolbar, a
really helpful development tool.

There is also a slight cleanup of the survey/urls.py to
stop the REGEX's being a bit too keen to match all urls.

Adding back in auth urls.

Dell XPS keyboard seems to like to add random characters sometimes. Need to add pre-commit hooks to run tests.

Add view function for generating a questionnaire

This commit has added the ability to render a json response
representing a form.