Create MakeFile to support local dev environment
To speed up the delivery and ability of the team to work on the eq platform, this commit creates a makefile
that wraps the docker toolchain (docker-machine, docker & docker-compose) to enable a dev or designer to
create a local dev environment that doesn't require any special local installations nor access to docker-hub
to rebuild an image.

To bring up a local environment simply run 'make' at a commandline in the root directory.